Enable isolated provider network metadata access
When an isolated provider network with no virtual routers metadata
access occurs in the qdhcp netns.

Without the force_metadata option in dhcp_agent.ini ns-metadata-proxy
is not enabled. ns-metdata-proxy sits in the ip netns and proxies
requests from 169.254.169.254 to the nova-api-metadata service
outside the netns.

This change adds the force_metadata option when
enable-isolated-metadata = True.

Related to LP Bug #1831935

Change-Id: I4fde7882be69772f168a1a1a201022bf9cf3cd06