Modify paths to ensure that charms' charmhelpers is first
The pre-patch version of unit-testing tested the git-repo version of
charm-helpers instead of the version that was synced to the charm.

The git-repo version of charmhelpers is pip installed to obtain the
testing part of charm-helpers, for the unit tests.  Whilst this is a bug
(really), it was compounded in that - due to how the paths worked - the
unit tests' git version of charm-helpers was in the path first, and thus
all unit tests were performed using that version of charm-helpers, and
not the version actually synced into the charm.  This has led to all
sorts of 'odd' double mocking of config in the tests as there are
potentially two copies of charm-helpers (as was the case for py27).  A
further commit will undo that madness.

This patchset moves the charm synced version of charm-helpers earlier
into the path so it gets picked first when the modules are loaded.

Change-Id: Iff0566c52a283ff0cd2fc7707d981d8c47f3e6b2