Add ovs-record-hostname to deferable service list
The ovs-record-hostname service was introduced in the openvswitch SRU
for bug #1915829, however this service did not make the deferable
services list for the neutron-gateway charm. This causes package
upgrades to restart the openvswitch-switch service. Add the
ovs-record-hostname to the deferable services list in order to prevent
unintended restarts of openvswitch-switch.

Closes-Bug: #1955498
Change-Id: I24a32f6f5a5c51b8b8ee62f88a973a126106fcd9